Establish a uniform naming convention
There is an overarching pattern when implementing the various coroutine
operations:
A. A type containing the coroutine state.
B. The implementation of Coroutine for the type in A.
C. A trait offering a single method which converts Self (maybe with
additional arguments) into the type in A. The trait may impose some
constraints on Self (usually Sized).
D. A blanket implementation of the trait in C for all types implementing
Coroutine which satisfy the constraints.

The naming convention is as follows:
A. The name of the type in A has the prefix `Co` followed by a verb
naming the operation the type models in camel case.
B. The name of the trait in C is just the verb part of the type in A
(i.e. the name of the type without the `Co` prefix).
C. The name of the single method of the trait in C is the snake case
version of the trait name.