Allow ContinuationSet to be implemented by anyone
Extend the definitions of ContinuationSet and State to allow for
throw-away implementations of ContinuationSet. These so-called
throw-away objects are just used to compute the type of the set of
continuations. See, for example, the `fsm::tests::MatchContinuation`
struct, which has no size, is never even instantiated and is only used
to prove that a chain of Match<(MyOut0, MyState0), Match<(MyOut1,
MyState1), ..>> can be used as the continuation variant (provided
MyState0, MyState1 etc. implement State).