Introduce `fsm::StateTransition` trait
The `StateTransition` trait is introduced to isolate an important
concert which until now was bundled with the `fsm::State` trait, and
that is to define the two possible outcomes of the transition:
- The state machine transitions into one of the possible next states,
while at the same time producing a corresponding output value.
- The state machine ends, producing a final value

This change improved ergonomics by eliminating an otherwise difficult to
avoid repetition when defining the type produced when the state machine
ends: formerly, `fsm::State::Exit`, now `<fsm::State::Transition as
StateTransition>::Exit`.